Tests/LibWeb/Text/expected/selection-auto-scroll.txt

before: 0
after: 80
hidden before: 0
hidden after: 0
